APIViewOrdering/views.py
from rest_framework import status
from rest_framework.decorators import permission_classes, api_view
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response


@api_view(["GET"])
@permission_classes([IsAuthenticated])
def view_with_api_view_first(request):
    return Response(status=status.HTTP_204_NO_CONTENT)


@permission_classes([IsAuthenticated])
@api_view(["GET"])
def view_with_permission_first(request):
    return Response(status=status.HTTP_204_NO_CONTENT)
