pyproject.toml
[tool.poetry]
name = "APIViewOrdering"
version = "0.1.0"
description = ""
authors = ["Florent <florent@nirlab.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
poetry-core = ">=1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
